# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/internal/campaigns/batch/budgets" : {
      "post" : {
        "tags" : [ "internal-campaign-controller" ],
        "operationId" : "getDailyBudgets",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "$ref" : "#/components/schemas/CampaignPacingDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CampaignPacingDTO" : {
        "required" : [ "advertiserId" ],
        "type" : "object",
        "properties" : {
          "dailyBudget" : {
            "type" : "number",
            "format" : "double"
          },
          "lifetimeBudget" : {
            "type" : "number",
            "format" : "double"
          },
          "advertiserId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
...
    }
  }
}